requirements.txt
torch==1.13.1
transformers==4.28.0
pytorch-metric-learning==2.1.2
faiss-gpu==1.7.2
# scipy==1.10.1
# numpy==1.26.4
loguru
tqdm